fixed FE

+++ content-service/routers/comments.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from content_queries import ContentQueries
 from auth import authenticator
 from model import CommentIn, CommentOut, EditComment
@@ -30,13 +30,12 @@ def delete_comment_by_id(
     content_queries: ContentQueries = Depends(),
     account: dict = Depends(authenticator.get_current_account_data)
 ):
-    return content_queries.delete_comment(id)
-    # get = content_queries.get_comment_by_id(id)
-    # creator = get['user_id']
-    # if account['id'] == creator:
-    #     return content_queries.delete_comment(id)
-    # else:
-    #     raise HTTPException(status_code=401, detail="not working")
+    get = content_queries.get_comment_by_id(id)
+    creator = get['user_id']
+    if account['id'] == creator:
+        return content_queries.delete_comment(id)
+    else:
+        raise HTTPException(status_code=401, detail="not working")
 
 
 @router.put('/api/comment/{id}', response_model=CommentOut)
@@ -46,10 +45,9 @@ def update_comment_by_id(
     content_queries: ContentQueries = Depends(),
     account: dict = Depends(authenticator.get_current_account_data)
 ):
-    return content_queries.edit_comment(id, body)
-    # get = content_queries.get_comment_by_id(id)
-    # creator = get['user_id']
-    # if account['id'] == creator:
-    #     return content_queries.edit_comment(id, body)
-    # else:
-    #     raise HTTPException(status_code=401, detail="not working")
+    get = content_queries.get_comment_by_id(id)
+    creator = get['user_id']
+    if account['id'] == creator:
+        return content_queries.edit_comment(id, body)
+    else:
+        raise HTTPException(status_code=401, detail="not working")
